+++ qa/rolling-upgrade/src/javaRestTest/java/org/elasticsearch/upgrades/TsdbIndexingRollingUpgradeIT.java
@@ -12,68 +12,37 @@ package org.elasticsearch.upgrades;
 import com.carrotsearch.randomizedtesting.annotations.Name;
 
 import org.elasticsearch.client.Request;
-import org.elasticsearch.client.Response;
 import org.elasticsearch.common.network.NetworkAddress;
-import org.elasticsearch.common.xcontent.XContentHelper;
 import org.elasticsearch.test.rest.ObjectPath;
-import org.elasticsearch.xcontent.XContentType;
 
-import java.io.IOException;
-import java.io.InputStream;
 import java.time.Instant;
 import java.util.Map;
 
-import static org.elasticsearch.upgrades.LogsIndexModeRollingUpgradeIT.enableLogsdbByDefault;
 import static org.elasticsearch.upgrades.LogsIndexModeRollingUpgradeIT.getWriteBackingIndex;
+import static org.elasticsearch.upgrades.LogsdbIndexingRollingUpgradeIT.getIndexSettingsWithDefaults;
+import static org.elasticsearch.upgrades.LogsdbIndexingRollingUpgradeIT.startTrial;
+import static org.elasticsearch.upgrades.TsdbIT.TEMPLATE;
 import static org.elasticsearch.upgrades.TsdbIT.formatInstant;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.greaterThan;
 import static org.hamcrest.Matchers.greaterThanOrEqualTo;
 import static org.hamcrest.Matchers.notNullValue;
 
-public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
+public class TsdbIndexingRollingUpgradeIT extends AbstractRollingUpgradeTestCase {
 
     static String BULK_ITEM_TEMPLATE =
         """
-            {"@timestamp": "$now", "host.name": "$host", "method": "$method", "ip": "$ip", "message": "$message", "length": $length, "factor": $factor}
+            {"@timestamp": "$now", "metricset": "pod", "k8s": {"pod": {"name": "$name", "uid":"$uid", "ip": "$ip", "network": {"tx": $tx, "rx": $rx}}}}
             """;
 
-    private static final String TEMPLATE = """
-        {
-            "mappings": {
-              "properties": {
-                "@timestamp" : {
-                  "type": "date"
-                },
-                "method": {
-                  "type": "keyword"
-                },
-                "message": {
-                  "type": "text"
-                },
-                "ip": {
-                  "type": "ip"
-                },
-                "length": {
-                  "type": "long"
-                },
-                "factor": {
-                  "type": "double"
-                }
-              }
-            }
-        }""";
-
-    public LogsdbIndexingIT(@Name("upgradedNodes") int upgradedNodes) {
+    public TsdbIndexingRollingUpgradeIT(@Name("upgradedNodes") int upgradedNodes) {
         super(upgradedNodes);
     }
 
     public void testIndexing() throws Exception {
-        String dataStreamName = "logs-bwc-test";
+        String dataStreamName = "k9s";
         if (isOldCluster()) {
             startTrial();
-            enableLogsdbByDefault();
-
             final String INDEX_TEMPLATE = """
                 {
                     "index_patterns": ["$PATTERN"],
@@ -81,7 +50,7 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
                     "data_stream": {
                     }
                 }""";
-            String templateName = "3";
+            String templateName = "2";
             var putIndexTemplateRequest = new Request("POST", "/_index_template/" + templateName);
             putIndexTemplateRequest.setJsonEntity(INDEX_TEMPLATE.replace("$TEMPLATE", TEMPLATE).replace("$PATTERN", dataStreamName));
             assertOK(client().performRequest(putIndexTemplateRequest));
@@ -91,7 +60,7 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
 
             String firstBackingIndex = getWriteBackingIndex(client(), dataStreamName, 0);
             var settings = (Map<?, ?>) getIndexSettingsWithDefaults(firstBackingIndex).get(firstBackingIndex);
-            assertThat(((Map<?, ?>) settings.get("settings")).get("index.mode"), equalTo("logsdb"));
+            assertThat(((Map<?, ?>) settings.get("settings")).get("index.mode"), equalTo("time_series"));
             assertThat(((Map<?, ?>) settings.get("defaults")).get("index.mapping.source.mode"), equalTo("SYNTHETIC"));
 
             ensureGreen(dataStreamName);
@@ -125,23 +94,21 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
         var bulkRequest = new Request("POST", "/" + dataStreamName + "/_bulk");
         StringBuilder requestBody = new StringBuilder();
         for (int i = 0; i < numDocs; i++) {
-            String hostName = "host" + i % 50; // Not realistic, but makes asserting search / query response easier.
-            String methodName = "method" + i % 5;
-            String ip = NetworkAddress.format(randomIp(true));
-            String message = randomAlphaOfLength(128);
-            long length = randomLong();
-            double factor = randomDouble();
+            String podName = "pod" + i % 5; // Not realistic, but makes asserting search / query response easier.
+            String podUid = randomUUID();
+            String podIp = NetworkAddress.format(randomIp(true));
+            long podTx = randomLong();
+            long podRx = randomLong();
 
             requestBody.append("{\"create\": {}}");
             requestBody.append('\n');
             requestBody.append(
                 BULK_ITEM_TEMPLATE.replace("$now", formatInstant(startTime))
-                    .replace("$host", hostName)
-                    .replace("$method", methodName)
-                    .replace("$ip", ip)
-                    .replace("$message", message)
-                    .replace("$length", Long.toString(length))
-                    .replace("$factor", Double.toString(factor))
+                    .replace("$name", podName)
+                    .replace("$uid", podUid)
+                    .replace("$ip", podIp)
+                    .replace("$tx", Long.toString(podTx))
+                    .replace("$rx", Long.toString(podRx))
             );
             requestBody.append('\n');
 
@@ -162,20 +129,20 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
             {
                 "size": 0,
                 "aggs": {
-                    "host_name": {
+                    "pod_name": {
                         "terms": {
-                            "field": "host.name",
+                            "field": "k8s.pod.name",
                             "order": { "_key": "asc" }
                         },
                         "aggs": {
-                            "max_length": {
+                            "max_tx": {
                                 "max": {
-                                    "field": "length"
+                                    "field": "k8s.pod.network.tx"
                                 }
                             },
-                            "max_factor": {
+                            "max_rx": {
                                 "max": {
-                                    "field": "factor"
+                                    "field": "k8s.pod.network.rx"
                                 }
                             }
                         }
@@ -189,13 +156,13 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
 
         Integer totalCount = ObjectPath.evaluate(responseBody, "hits.total.value");
         assertThat(totalCount, greaterThanOrEqualTo(4096));
-        String key = ObjectPath.evaluate(responseBody, "aggregations.host_name.buckets.0.key");
-        assertThat(key, equalTo("host0"));
-        Integer docCount = ObjectPath.evaluate(responseBody, "aggregations.host_name.buckets.0.doc_count");
+        String key = ObjectPath.evaluate(responseBody, "aggregations.pod_name.buckets.0.key");
+        assertThat(key, equalTo("pod0"));
+        Integer docCount = ObjectPath.evaluate(responseBody, "aggregations.pod_name.buckets.0.doc_count");
         assertThat(docCount, greaterThan(0));
-        Double maxTx = ObjectPath.evaluate(responseBody, "aggregations.host_name.buckets.0.max_length.value");
+        Double maxTx = ObjectPath.evaluate(responseBody, "aggregations.pod_name.buckets.0.max_tx.value");
         assertThat(maxTx, notNullValue());
-        Double maxRx = ObjectPath.evaluate(responseBody, "aggregations.host_name.buckets.0.max_factor.value");
+        Double maxRx = ObjectPath.evaluate(responseBody, "aggregations.pod_name.buckets.0.max_rx.value");
         assertThat(maxRx, notNullValue());
     }
 
@@ -204,7 +171,7 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
         queryRequest.addParameter("pretty", "true");
         queryRequest.setJsonEntity("""
             {
-                "query": "FROM $ds | STATS max(length), max(factor) BY host.name | SORT host.name | LIMIT 5"
+                "query": "FROM $ds | STATS max(k8s.pod.network.rx), max(k8s.pod.network.tx) BY k8s.pod.name | SORT k8s.pod.name | LIMIT 5"
             }
             """.replace("$ds", dataStreamName));
         var response = client().performRequest(queryRequest);
@@ -214,36 +181,16 @@ public class LogsdbIndexingIT extends AbstractRollingUpgradeTestCase {
         String column1 = ObjectPath.evaluate(responseBody, "columns.0.name");
         String column2 = ObjectPath.evaluate(responseBody, "columns.1.name");
         String column3 = ObjectPath.evaluate(responseBody, "columns.2.name");
-        assertThat(column1, equalTo("max(length)"));
-        assertThat(column2, equalTo("max(factor)"));
-        assertThat(column3, equalTo("host.name"));
+        assertThat(column1, equalTo("max(k8s.pod.network.rx)"));
+        assertThat(column2, equalTo("max(k8s.pod.network.tx)"));
+        assertThat(column3, equalTo("k8s.pod.name"));
 
         String key = ObjectPath.evaluate(responseBody, "values.0.2");
-        assertThat(key, equalTo("host0"));
+        assertThat(key, equalTo("pod0"));
         Long maxRx = ObjectPath.evaluate(responseBody, "values.0.0");
         assertThat(maxRx, notNullValue());
-        Double maxTx = ObjectPath.evaluate(responseBody, "values.0.1");
+        Long maxTx = ObjectPath.evaluate(responseBody, "values.0.1");
         assertThat(maxTx, notNullValue());
-    }
-
-    protected static void startTrial() throws IOException {
-        Request startTrial = new Request("POST", "/_license/start_trial");
-        startTrial.addParameter("acknowledge", "true");
-        assertOK(client().performRequest(startTrial));
-    }
-
-    static Map<String, Object> getIndexSettingsWithDefaults(String index) throws IOException {
-        Request request = new Request("GET", "/" + index + "/_settings");
-        request.addParameter("flat_settings", "true");
-        request.addParameter("include_defaults", "true");
-        Response response = client().performRequest(request);
-        try (InputStream is = response.getEntity().getContent()) {
-            return XContentHelper.convertToMap(
-                XContentType.fromMediaType(response.getEntity().getContentType().getValue()).xContent(),
-                is,
-                true
-            );
-        }
     }
 
 }
